implemented tooltips
Tooltips will show when a field with any warnings or errors is hovered
or if any of its labels (text to its left, or a status detail) are clicked.

Fields with warnings and errors will also change border color if hovered or
focused to better show this.

This will need to be updated to process tooltips for cards with quantities in
excess of 4 and unrecognized lines are properly implemented for the status box.

**NOTE:** jQuery UI updated to 1.11.2; added tooltip.js