Clean up generatecards.sh
`set -e` will stop execution if errors are encountered. (For instance,
if the script is run on a Windows environment, `/usr/bin/env python3`
may not be set and will cause the Python script call to fail)

File descriptors starting with a glob, such as `*.js` are changed to be
preceded with `--` to ensure strangely-named files are not improperly
interpreted as flags.